aws-ecs-1: add configurable logging drivers
The new ecs.logging-drivers setting controls the list of logging drivers
available for use with ECS.  The defaults are "json-file", "awslogs", and
"none".  Other logging drivers (such as those installed via a plugin) can be
enabled.  The default set of logging drivers can be reduced or removed to
implement policy for allowable logging drivers in a cluster.