Refactor default settings files to allow sharing
Previously, storewolf would read two hard-coded paths, a global defaults.toml
and a variant-specific override-defaults.toml.  TOML doesn't have a way of
including other files, so the layout was fairly strict.  Not all variants want
all of the settings in defaults.toml, and there's no way for a variant to
remove the global settings defined in defaults.toml, only add or replace, which
can be a problem if, for example, it doesn't want the default host containers.

This change makes each variant have a `defaults.d` directory in its model
where it can include any number of files containing default settings.  In this
way, common settings can be shared through symlinks, and settings that are not
common to all variants can be moved to separate files and not included by a
variant.