models build.rs: remove early exit for target toolchain
This is a partial revert of 79465cd6d3f1e6f121dd15915a0346fee2bbe9ac; the nicer
structure for main() is kept, but the actual early-exit is removed.

We don't actually need the early-exit if it's safe to run this build.rs in
parallel, which it is since 3943a32212a424c6dc0c0c97e3f15b0f627e937c introduced
safe link swapping.  The other action it takes, README generation, is skipped
during production builds, and would be obvious if it went wrong during local
builds due to version controlled changes, though it should also be safe there
as long as our READMEs don't balloon and become unsafe to write in parallel.

The reason for removing this is that it effectively worked by chance, because
we were running non-static builds before static builds.  We'd like to run both
sets of builds in parallel to save a significant amount of time.  Static
builds, as run in os.spec, only run build.rs once, with a vendor of
"bottlerocket".  This means build.rs was skipped, and unless non-static builds
won the race and ran their build.rs before static builds got to that crate, the
link wouldn't be created and the build would fail.