ecs: add additional API configurations
With this, 4 additional configurations for the ECS agent are supported
though the API.

There are two configuration files used to set up the ECS agent:

- /etc/ecs/ecs.config.json
- /etc/ecs/ecs.config

We favor the former to add new configurations, and we only use the latter
on special cases, i.e. when the configurations to be added aren't
modeled as part of the struct that represents the agent's configuration,
or when special deserialization is used to parse the configurations.

The configurations added in this change are as follows:

ECS_CONTAINER_STOP_TIMEOUT: supported through the container-stop-timeout
API; this configuration is rendered in the /etc/ecs/ecs.config file since
this configuration is of type Duration (1m, 1s, 1h). This type must be
parsed by calling the time.ParseDuration function which isn't called
under the hood by the serialization libraries used in the ECS agent.

ECS_ENGINE_TASK_CLEANUP_WAIT_DURATION: supported through the
task-cleanup-wait API; this configuration is of the same type as the
previous configuration and was rendered following the same reasoning.

ECS_RESERVED_MEMORY: supported through the reserved-memory API; this
configuration is rendered in /etc/ecs/config.ecs.json since the
configuration's type can be deserialized without additional helper
functions.

ECS_TASK_METADATA_RPS_LIMIT: this configuration represents a
comma-separated string with two values used to set the throttling rates
in the metadata service exposed by the ECS agent. These values don't
have to be set together, since the ECS agent will use default values if
either is missing. Thus, this configuration is supported through the
metadata-service-rps and metadata-service-burst APIs. Both
configurations are rendered in the /etc/ecs/confing.ecs.json file, since
the configurations' type can be deserialized without additional helper
functions.

Signed-off-by: Arnaldo Garcia Rincon <agarrcia@amazon.com>