apiserver: Add /report endpoint for CIS reporting
This adds a new `/report` endpoint. GETs to this endpoint get a list of
the available reports which can then be called directly. Initially this
is only `/report/cis` that can be used to trigger a bloodhound CIS
benchmark report, returning the results. The type argument controls
which report to run. Right now there is only the CIS report, but this
will be expanded in the future as we add other compliance or benchmark
reports.

```
/report
[{"name":"cis","description":"CIS Bottlerocket Benchmark"}]
```

The bloodhound options for `--level` and `--format` are supported with:

```
/report/cis&level=2&format=json"
```

Signed-off-by: Sean McGinnis <stmcg@amazon.com>