Fix `sources/models` README from generated content
New k8s-1.26 variants were recently added, but the README.md file was
modified directly with the updated variants. This would normally be
fine, but like many README.md files in the repo, this file is generated
based on code comments in another file. This would result in the file
showing modified whenever doing a local build.

This updates the code comments so the generated README has all necessary
variant information.

This also cleans up some changes in title level that have been
copy/pasted to a few different variants. Title levels are now the same
for all variants.

There is also a problem that GitHub does not work with hyperlinks to
symlinks. So as we update new variants, we point the old variant to the
new variant if the model does not change. This causes the majority of
the hyperlinks in the README on GitHub to go to 404 error pages.

This updates the link references to go to the "real" file in the repo so
all hyperlinks lead to a valid destination that shows the model being
used for that variant.

Signed-off-by: Sean McGinnis <stmcg@amazon.com>