feat: add crates for settings plugins
Settings plugins are an intermediate step toward the planned world of
settings extensions. Variants can define a monolithic settings struct
that combines custom and shared types, which can be loaded at runtime
by Bottlerocket's existing API server and datastore implementations.

Since Rust lacks a stable ABI, the key to making this work is for the
host program (the API server or datastore creator) to load the plugin
as a cdylib that implements a particular FFI contract. The abi_stable
crate does the heavy lifting here with its support for load-time type
checking and FFI-safe wrapper types.

Plugins must define a struct and implement serialize, deserialize,
and default methods for that struct. That allows the host program to
instantiate a new version of the struct, make increment changes, save
the result to the datastore, and load it again, all without compile-
time access to the type. That in turn means the host programs can be
built once and support arbitrary settings at runtime.

Signed-off-by: Ben Cressey <bcressey@amazon.com>